Add files via upload
Things That I Added: Registration Form, AdminMenu (For Admins), SelectionMenuController(For Users)
Things That I Updated: Deleting an Appointment on the User side now takes you to another form, which asks you to type in the user id# and name associated with appointment you want to delete, Hitting the Reschedule button opens up a popup box, which asks you to type in the id# and name of appointment you want to update. When you click confirm, it asks you to select available appointments (If they have any) associated with the contact name you scheduled with. You can now only view tables from the Admin Side and not the User side (Makes sense because technically users should not be able to see everyones information.